Add framework for floorplan editor
The floorplan will be made using svg (and probably the interface
aswell) which means it won't have to be specially serialized to
just print the floorplan, etc.

The system is going to be mode based, where there's an edit mode,
furnature mode, view mode, etc. which should make it easier to build
because each system can be independant of the rest reducing complexity.

The framework isn't set in stone yet though, I believe one major
change I'll introduce is a state namespace for each handler so it
doesn't have to think about which names other handlers are using.