Entirely rewrite floorplan editor
The previous interface used coordinates relative to many different
things (the screen, the document, absolute, relative, etc. etc.)
which was the main source of problems.

The new interface uses a single type of coordinate: An absolute
position in the SVG document (which can be used as a relative offset
too).  It also uses the svg.js library (aswell as the svg.panzoom.js
plugin for it) to help with managing the SVG. Introducing third-party
libraries isn't something I do lightly, but I think it is worth it.