Add backend layer to floorplan editor
This interface will manage communication with the server and try
and stay out of the way.

It keeps a set of diffs and will update the server as necessary.
The editor was reworked to simply push changes to the backend, then
pull them back again to make changes to the document, eliminating
any difference between data fetched from the server and data pushed
from the editor.

Currently it doesn't communicate with the server, but it's built
to have that functionality added in the future so it shouldn't be
too difficult.