Add unit separate and combine Units methods
The separate method takes a number and a unit system and iterates
over the units in the system from biggest to smallest, converting
as much of the number to the current unit as possible while still
being a whole number.

The combine method does the opposite, reconstructing the original
number.