Add Precise mode to editor
This is the start of the actual interface I want the editor to have.

It's going to require some vector math, so I'v included the
threejs-math library, and while the whole thing is pretty big, I'm
only using Vector2 from it which isn't too bad.

It also required some changes to the editor:

- Rename mapPoints to mapSelected,
- and add a new method mapPoints which allows you to specify the points to map
- Add ui group to draw
- Have keydown & keyup events listen on the document