Add separate file for geometric math
Mostly left the existing math from main.js where it is (to be adapted
and moved over later), but extended SVG.Line to support the inside()
method aswell as a method called closestPoint which just gives you
the point on the line where the given point is, if any.

These will be necessary for future additions to the floorplan editor.