Add reverse Pointmap lookup table in backend
This is something I was going to leave for later as, while grossly
inefficient, the current method was fast enough for my needs. Now
however I'm going to need to do a ton of reverse lookups in the
near future and I think this will just make it easier in addition
to being more efficient. (Although I suppose I could've written a
method that works like whichPointMap but searches each time.)