Add checks and options to point removal in the backend
The removePoint method now does checks to ensure that no pointmap
references the point before deletion. It can also delete all pointmaps
that reference it if given the 'recurse' or 'unmap' option, and
will furthermore delete all orphaned points if given the 'recurse'
option.

Orphaned points is a new concept that means any point not connected
by pointmaps to the origin point. The origin point is the first
point in the diffs that currently exists.