Add handler for removing points
This required a little hack to support alongside the point adding
handler but turned out okay. I have some ideas for how to improve
the whole system in the future to maybe make this easier.

So really what happened was two new handlers were added, and edit
handler (which so far only supports removing points), and a main
point handler which handles executing either add or edit depending
on the state of things.