Add very basic furniture support
All you can do is place furniture by clicking on an empty space in
the floorplan, and delete it by pressing delete in the selection
menu, but I think it's got all the boiler-plate done and is ready
to have the nuts and bolts added.

This also fixes a bug in history.updateIds where the id wouldn't
be updated for pointmaps, I believe.