Update object ID system
This adds support for the new system-unique string IDs from the
server and furthermore always uses local IDs to reference objects,
and only maps to and from those IDs only when communicating with
the server. This should make things less error-prone because there's
only one place where IDs need to be updated, instead of everywhere
throughout the program's memory.