Fix backend history system
Instead of having groups of diffs encompassing diff's A to B, there
are now diff marks, which is just an array of diff indexes. They're
used the same way in essence, but being just a mark means they don't
need a last group member index which keeps things more simple.

The diffs now always records the operation as new, replace, or
remove, which makes updating the server a little easier and less
error prone. (One issue was that the server would be given a 'new'
patch for the same object multiple times instead of 'new' followed
by 'replace'.)