Fix backend and editor state mismatch
When reconstructTo was called on the backend, it's state was
reconstructed to what it looked like at diff position D. Afterwords
the editor might—or in the case of this program, would—step through
history much like the backend applying differences in order to
reconstruct it's state to where it needed to be. The issue came
when it needed to reference some data that wasn't in the diff, for
example points a pointmap refers to, and there was a mismatch between
what the data should've looked like at that point for the editor
and what it actually was in the backend.

One way of solving this would be to have the editor store it's own
state of things and push and pull differences to and from the
backend, but then what's the point of the backend? So now you can
provide the backend with a callback function for applying differences.
This actually simplifies quite a few things that depend on the
backend aswell.

Now the state of the editor is always the same as the state of the
backend.