Fix event handlers
I misunderstood the preventDefault method for events thinking that
it would prevent other handlers from executing, but actually
stopPropagation or stopImmediatePropagation must be called for that,
so in large part this change is implementing the handled function
which calls both preventDefault and stopImmediatePropagation.

In addition there were some minor updates to the logic in a few
handlers aswell as modifying the furniture handler to use mousedown
and mouseup instead of dblclick so precisePointHandler can properly
stop propagation.