Update styles, mainly for the landing page
Added a few paragraphs of possible selling points and prominent
signup buttons. The login button used to be an icon, but now it's
text as the icon wasn't conveying the purpose of the button easily
enough in my opinion.

I'v added two classes for anchor tags: button and pop.  The button
class styles the element as a button and the pop tag gives the link
high contrast colors (and is really only meant to be used in concert
with the button class).