Add history compression
Now, if there's a diff for the same object without a mark inbetween,
the old diff's value is set to the new value and no new diff is
created. It could be taken further by deleting a "new" operation
if there is a "remove" operation afterwords, but for now this should
be a large improvement.

One caveat of the current system is that each time this is done
(which is all the time in this program) the history truncated value
is updated causing the server to receive the full floorplan again.
This isn't such a big deal, but perhaps history should also store
the position that it was truncated at so allow delta updates whenever
possible.