Add point alignment snapping
Now when moving a point it's x and y axis are snapped independently
to all other points if their axis is close enough.  Currently if
points are behind walls they are still snapped to, which I consider
a bug but I'm not going to worry about it now.

This also adds the ability to disable all snapping by pressing
shift.