Change how furniture editing works
The furniture menu is now opened when furniture is selected, there
is no longer a cancel button, and escaping does not cancel.

I'm removing cancel in part because it wasn't working and I didn't
want to fix it, and also because I didn't think added enough value
to warrent it's existence.