Put head in html tag and add viewport tags
I'v added the viewport tag to all html documents, in most it's
simply setting content="width=device-width" excepting the editor
which has a bit more to it.

Most of the heads were at the root level of the HTML document which
I remedied in addition to the missing body tags in many documents.