Don't fire select event if selection wouldn't change
This is mostly for a future change where the user will click on a
point, then have the option to select whether they wan't to move
it or not, and finally click and drag to either move the point or
create a new one. Without this change it would end up resetting
their choice when they go to drag the point.

Although it's not bad behavior in my opinion, I can see some use
cases where you would want an event to be fired again.