Make use on touch-devices tolerable
Since I want the mouse and touch interfaces to be as similar as
possible (both to make it easier to program and easier for users)
this required changing the mouse interface aswell, which I wanted
to do anyway.

It's still not perfect, but now it only makes use of double click
once and does away with long presses for now. The only real issue
with use on a touch device is it can be hard to select walls and
points since they're so small on small screens. This is a problem
for the future.