Increase the size of points and pointmaps
This is to try and make it easier to press them on touch devices,
but it's still not always easy. May either have to scale with
viewport zoom which could be problematic or else have padding around
them so that when you're close, even if you're not on them, it will
register.