Fix grid display
Add a margin to the grid because something is probably wrong with
my math and when the aspect ratio is close to 1 there is areas of
the screen with no grid.

Also hide the grid until the floorplan data is loaded so it doesn't
look so buggy when the viewbox is resized.