Update backend mapPoints method
It now takes a parameters argument along with the ID of the map to
update like furniture maps. It uses a new method updatedObject to
validate parameters and build to new object. In the future I want
all these functions function the same way using this updatedObject
method.

While it should have been a separate commit, the precisePointHandler
function was also updated so fix some issues with the snapmap stuff,
which led to the changes described above.