Fix email verification warning
It behaved differently if you had an no verified email and a verified
email different then the setting. Now it behaves the same no matter
what. Importantly, you can now access the verification page in
either scenario, which is good because there's no other way to get
to the email verification page.