Fix directly deleting cache data in furniture menu
The backend.obj method returns a mutable pointer to an object in
the cache, which can be helpful if you want to keep values updated
but really bad if you forget and mutate it. I'll have to look into
a solution for this. Maybe you can return an unmutable variable in
which case that's what I'll do. Otherwise make it a copy by default
and add a separate method to give a pointer for when you need it.