Use pointmaps endpoint to validate pointmap types
The pointmaps endpoint was recently added to the API and it returns
a list of valid pointmap types. Now we just use that endpoint to
verify valid types instead of hardcoding them in here.

This also brought about some changes to the backend and editor
initialization process: Both set the initialization field to a
promise that resolves only when they're fully initialized. This
prevents race conditions that would otherwise occur. Furthermore
furniture parameters are now fetched in the backend instead of
(strangely) from the main.js file.