Add furniture duplicate button to furniture menu
I'v added a new section to the furniture menu for operations like
this, and what was originally only part of the furniture menu is
now the furniture parameters menu.

I'v also taken the opportunity to refactor it a bit, mostly by
splitting up some of the original furniture menu function into a
couple others.