Add multi-selection
Multi-selection is going to be very important for moving multiple
points, etc. At the moment it's still fairly limited and only useful
for saving a bit of time when deleting a few different things or
changing pointmap types etc., but I'll rework various handlers to
support moving multiple objects as a group in the future (see
thoughts below).

In this selection mode I want to have press, drag, and release
create a box from the press to the release which selects everything
inside of it.

I'm thinking that while it sounds a little tedious it might be a
good idea to have changes to the SVG elements themselves push diffs
to the backend. I think this would end up making all this front-end
code—and I can't overstate this enough—a lot easier.  For instance
moving or rotating multiple objects would just mean putting them
in a group, transforming them, then have svg.js take them out of a
group and apply the transform to each individually. Of course there
are some issues I see with this already but I still think it might
be a good idea.