Add loading feedback to editor
At various points during the loading process a status update is
issued using the new ui.wait() function to give an idea of the
different stages and show that things are happening.  This is mostly
for slow network connections when the scripts aren't cached yet.

Also add a <noscript> error message.