Prevent multiple inputs on most pointer handlers
This is more inline with the desired behavior anyway, but it's
extremely important for the next change.

It's implemented by creating a single handler function which will
stop propagation of an event if multiple inputs are detected, so
if you have some handlers that accept multiple inputs you can just
place this handler after it.

One caveat of the implementation at the moment is that it doesn't
handle mouse inputs as you would expect (due to the design of pointer
events).  This could be fixed in the future without much trouble
if it's a big issue, if it's even the desired behavior.