Make points and pointmaps smaller
This should allow greater precision in the floorplans, yet I think
it's still not too small. It may be adjusted a bit in the future
(before people have made many floor plans).

This has probably caused a few issues, for example door swing by
pointer is now more difficult. The solution is to adapt all those
handlers to use the selection instead of checking .inside themselves
or whatever, because the selection selects even objects your not
directly over.