Make pointer door swing updates live
Now the door swing is updated on every pointermove (which I suppose
I should throttle), and also uses params.threshold (650) like
everything else instead of it's own value of 500.

One thing you'll note about the behavior is that the hinge isn't
updated on pointermove, I don't see this as an issue and it might
even be the desired behavior so I didn't change that.