Add long press on canvas to add furniture
Not sure if I'll stick with this, but I need to either reduce the
number of items in the toolbar or add dropdown menus to it.  I don't
think I'll remove the furniture adding button though, because that's
the best way to document other ways of doing the same thing.