Put some tools in a dropdown menu
The toolbar was getting far too big and cluttered. I'm not entirely
in love with how I'v done this, but it is an improvement.  I looked
at some libraries (with w2ui chief among them) to do toolbars, but
they didn't seem to be able to do what I wanted. I'll look into it
more later.

The selection bit has to be improved opon too.