Fix bug in id mapping
When remapping, the old mapping wasn't removed, it's just a new one
was added. This might fix an issue I'v had with pushing invalid IDs
but I couldn't figure out how to reproduce it so no way of really
testing except waiting to see if it happens again (it wasn't too
infrequent.)

If it does stop, I expect the issue may have ocurred when one server
id was unmapped and the other was left, so on a patch it was used
to remap ids to. I'd have to look into it more to see if this was
possible.